feat(front-kit-3): moved product card
