+++ examples/shop-with-blog/client/src/pages/shop/Category/Category.js
@@ -8,12 +8,11 @@ import {
   CategoryLayout,
   CategoryArea,
   ShowingOutOf,
-  ProductList,
   ShowMore,
   Loader
 } from '@deity/falcon-ecommerce-uikit';
 import { SearchConsumer, SortOrderPickerProvider, getFiltersData } from '@deity/falcon-front-kit';
-import { Sidebar, Responsive, SortOrderPicker, FiltersSummary } from '@deity/falcon-ui-kit';
+import { Sidebar, Responsive, SortOrderPicker, FiltersSummary, ProductList } from '@deity/falcon-ui-kit';
 import { Filters } from './Filters';
 
 const copy = item => item && JSON.parse(JSON.stringify(item));
@@ -78,7 +77,7 @@ const CategoryPage = ({ id }) => (
               )}
               <Box gridArea={CategoryArea.content}>
                 <FiltersSummary data={filtersData} />
-                <ProductList products={items} />
+                <ProductList items={items} />
               </Box>
               <FlexLayout gridArea={CategoryArea.footer} flexDirection="column" alignItems="center">
                 {pagination.nextPage && <Divider />}
