--- packages/falcon-ecommerce-uikit/src/ProductList/index.tsx
@@ -1,1 +1,0 @@
-export * from './ProductList';
